refactor(*): add lints and eliminate warnings

+++ taos-optin/src/tmq/raw.rs
@@ -328,7 +328,7 @@ pub(super) mod tmq {
                 vgroup_id
             );
 
-            if tmq_resp.0 as i32 > 0 {
+            if tmq_resp.0 > 0 {
                 Ok(tmq_resp.0 as _)
             } else {
                 let err_str = self.err_as_str(tmq_resp);
